shailesh_blog/requirements1.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
django-taggit==4.0.0
pytz==2021.1
sqlparse==0.4.1
typing_extensions==4.7.1
